Data.IORef.Extra

Description

This module extends Data.IORef with operations forcing the value written to the IORef. Some of these functions are available in later versions of GHC, but not all.

Documentation

module Data.IORef

writeIORef' :: IORef a -> a -> IO ()

Evaluates the value before calling writeIORef.

atomicWriteIORef' :: IORef a -> a -> IO ()

Evaluates the value before calling atomicWriteIORef.

atomicModifyIORef_ :: IORef a -> (a -> a) -> IO ()

Variant of atomicModifyIORef which ignores the return value

atomicModifyIORef'_ :: IORef a -> (a -> a) -> IO ()

Variant of atomicModifyIORef' which ignores the return value