System.Environment.Extra

Description

Extra functions for System.Environment. All these functions are available in later GHC versions, but this code works all the way back to GHC 7.2.

Documentation

module System.Environment

getExecutablePath :: IO FilePath

Returns the absolute pathname of the current executable.

Note that for scripts and interactive sessions, this is the path to the interpreter (e.g. ghci.)

Since: 4.6.0.0

lookupEnv :: String -> IO (Maybe String)

Return the value of the environment variable var, or Nothing if there is no such value.

For POSIX users, this is equivalent to getEnv.

Since: 4.6.0.0